Movie_API/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.1
django-cors-headers==3.10.1
django-mssql-backend==2.8.1
django-pandas==0.6.6
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.22.0
openapi-codec==1.3.2
packaging==21.3
pandas==1.3.5
PyJWT==2.3.0
pyodbc==4.0.32
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
templates==0.0.5
threadpoolctl==3.0.0
uritemplate==4.1.1
urllib3==1.26.7
